Add rofi theming

+++ .config/rofi/powermenu/powermenu.sh
@@ -15,19 +15,19 @@
 # full_circle     full_square     full_rounded     full_alt
 # row_circle      row_square      row_rounded      row_alt
 
-theme="drop_square"
+theme="full_circle"
 dir="$HOME/.config/rofi/powermenu"
 
 # random colors
-styles="row_square"
-color="bluish.rasi"
+styles=($(ls -p --hide="colors.rasi" $dir/styles))
+color="${styles[$(( $RANDOM % 8 ))]}"
 
 # comment this line to disable random colors
-#sed -i -e "s/@import .*/@import \"$color\"/g" $dir/styles/colors.rasi
+sed -i -e "s/@import .*/@import \"$color\"/g" $dir/styles/colors.rasi
 
 # comment these lines to disable random style
-#themes=($(ls -p --hide="powermenu.sh" --hide="styles" --hide="confirm.rasi" --hide="message.rasi" $dir))
-#theme="${themes[$(( $RANDOM % 24 ))]}"
+# themes=($(ls -p --hide="powermenu.sh" --hide="styles" --hide="confirm.rasi" --hide="message.rasi" $dir))
+# theme="${themes[$(( $RANDOM % 24 ))]}"
 
 uptime=$(uptime -p | sed -e 's/up //g')
 
@@ -42,11 +42,7 @@ logout=""
 
 # Confirmation
 confirm_exit() {
-	rofi -dmenu\
-		-i\
-		-no-fixed-num-lines\
-		-p "Are You Sure? : "\
-		-theme $dir/confirm.rasi
+    echo "yes"
 }
 
 # Message
@@ -55,27 +51,65 @@ msg() {
 }
 
 # Variable passed to rofi
-options="$lock\n$logout\n$suspend\n$reboot\n$shutdown"
+options="$shutdown\n$reboot\n$lock\n$suspend\n$logout"
 
-chosen="$(echo -e "$options" | $rofi_command -p "Uptime: $uptime" -dmenu -selected-row 0)"
+chosen="$(echo -e "$options" | $rofi_command -p "Uptime: $uptime" -dmenu -selected-row 2)"
 case $chosen in
     $shutdown)
-        systemctl poweroff
+		ans=$(confirm_exit &)
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+			systemctl poweroff
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
         ;;
     $reboot)
-        systemctl reboot
-        ;;
-    $suspend)
-        mpc -q pause
-        systemctl suspend
-        ;;
-    $logout)
-        mpc -q pause
-        killall dwm
-        pkill -KILL -u $USER
+		ans=$(confirm_exit &)
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+			systemctl reboot
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
         ;;
     $lock)
-        xsecurelock
+		if [[ -f /usr/bin/i3lock ]]; then
+			i3lock -e -f -c 000000 --verif-text="" --wrong-text="WRONG"
+		elif [[ -f /usr/bin/betterlockscreen ]]; then
+			betterlockscreen -l
+		fi
+        ;;
+    $suspend)
+		ans=$(confirm_exit &)
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+			mpc -q pause
+			amixer set Master mute
+			systemctl suspend
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
+        ;;
+    $logout)
+		ans=$(confirm_exit &)
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+			if [[ "$DESKTOP_SESSION" == "Openbox" ]]; then
+				openbox --exit
+			elif [[ "$DESKTOP_SESSION" == "bspwm" ]]; then
+				bspc quit
+			elif [[ "$DESKTOP_SESSION" == "i3" ]]; then
+				i3-msg exit
+            else
+                loginctl kill-session $XDG_SESSION_ID
+			fi
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
         ;;
 esac
-
